backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-jose[cryptography]
httpx
neo4j